+++ crates/config/src/validate.rs
@@ -132,7 +132,7 @@ pub fn in_range<T: Bounds + 'static, D, C>(min: T, max: T) -> Validator<T, D, C>
     Box::new(move |value, _, _| r::range::apply(value, (&min, &max)).map_err(map_err))
 }
 
-/// Validate an extends value is either a file path or secure URL.
+/// Validate an `extend` value is either a file path or secure URL.
 pub fn extends_string<D, C>(value: &str, _data: &D, _context: &C) -> Result<(), ValidateError> {
     let is_file = is_file_like(value);
     let is_url = is_url_like(value);
@@ -156,7 +156,7 @@ pub fn extends_string<D, C>(value: &str, _data: &D, _context: &C) -> Result<(), 
     Ok(())
 }
 
-/// Validate a list of extends values are either a file path or secure URL.
+/// Validate a list of `extend` values are either a file path or secure URL.
 pub fn extends_list<D, C>(values: &[String], data: &D, context: &C) -> Result<(), ValidateError> {
     for (i, value) in values.iter().enumerate() {
         if let Err(mut error) = extends_string(value, data, context) {
@@ -169,7 +169,7 @@ pub fn extends_list<D, C>(values: &[String], data: &D, context: &C) -> Result<()
     Ok(())
 }
 
-/// Validate an extends value(s) is either a file path or secure URL.
+/// Validate an `extend` value is either a file path or secure URL.
 pub fn extends_from<D, C>(value: &ExtendsFrom, data: &D, context: &C) -> Result<(), ValidateError> {
     match value {
         ExtendsFrom::String(string) => extends_string(string, data, context)?,