+++ crates/types/src/strings.rs
@@ -1,6 +1,7 @@
 use crate::{SchemaType, Schematic};
 use std::net::{Ipv4Addr, Ipv6Addr};
 use std::path::{Path, PathBuf};
+use std::time::{Duration, SystemTime};
 
 #[derive(Clone, Debug, Default)]
 pub struct StringType {
@@ -55,3 +56,6 @@ impl_string_format!(PathBuf, "path");
 
 impl_string_format!(Ipv4Addr, "ipv4");
 impl_string_format!(Ipv6Addr, "ipv6");
+
+impl_string_format!(SystemTime, "time");
+impl_string_format!(Duration, "duration");